+++ CHANGELOG.md
@@ -11,14 +11,17 @@ with the exception that all versions are currently `0.x.x` and may include break
 
 ### Added
 
-- Use response headers to determine whether to parse HTTP responses into JSON
 - Added possibility for additional headers and proxies (handled by requests lib)
-- get() can also return `list` type as JSON can also be a list
 - Documentation for all functions
+
+### Breaking Changes
+
+- Renamed get_blogs_time() to get_blogs_for_year_month()
 
 ### Changed
 
-`N/A`
+- Use response headers to determine whether to parse HTTP responses into JSON
+- get() can also return `list` type as JSON can also be a list
 
 ### Fixed